services/airflow/pyproject.toml
[tool.poetry]
name = "bzd-data-pipelines"
version = "0.0.1"
description = "Pipelines to organize products uploads at BerlinzuDir"
authors = ["Jakob J. Kolb <jakob.j.kolb@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.10"
returns = "^0.16.0"
fsspec = "^2021.7.0"
polars = "0.9.12"
api-wrappers = {path = '../../libraries/api-wrappers', develop=true}
paramiko = "^2.8.1"
apache-airflow = "^2.2.2"
ramda = "^0.6.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = { version = "*", allow-prereleases = true } # according to https://github.com/peterjc/flake8-black#python-package-management
flake8-black = "^0.2.1"
pytest = "^6.2.4"
testfixtures = "^6.17.1"
pytest-xdist = "^2.3.0"
mypy = "^0.910"
types-requests = "^2.25.0"
pytest-dotenv = "^0.5.2"
types-Markdown = "^3.3.0"
types-PyMySQL = "^1.0.0"
types-croniter = "^0.1.4"
types-python-dateutil = "^0.1.4"
types-setuptools = "^57.0.0"
types-tabulate = "^0.1.1"
types-toml = "^0.1.3"
responses = "^0.14.0"
vcrpy = "^4.1.1"
mock = "^4.0.3"

[tool.pytest]
env_files =".env"
looponfailroots="dags"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
